move non-page file out of ‘pages’
